Refactor SlackHelper for use in specs
